+++ CHANGES.md
@@ -5,6 +5,7 @@
   * Store reflection objects separate from it's definition.
   * [BC BREAK] Change `registerFactory` arguments order.
   * New `has` method.
+  * Factory arguments can have literals.
 
 ### 0.1.0 (2014-10-03)